+++ src/modules/tasks/ModelsTypes.ts
@@ -1,6 +1,6 @@
 import { FileT, PermissionsT } from '@/modules/common/CommonTypes';
 
-export enum ModelCategory {
+enum ModelCategory {
     simple = 'MODEL_SIMPLE',
     head = 'MODEL_HEAD',
 }
@@ -10,7 +10,7 @@ export type ModelT = {
     category: ModelCategory;
     compute_task_key: string;
     address?: FileT;
-    permissions: PermissionsT;
+    permissions?: PermissionsT;
     owner: string;
     creation_date: string;
 };